refactored routes
