+++ greg_magdsick/lib/server.js
@@ -1,9 +1,13 @@
 const app = require('express')();
-const router = require(__dirname + '/../routes/router');
+const pedal = require(__dirname + '/../routes/pedal');
+const motor = require(__dirname + '/../routes/motor');
+const fast = require(__dirname + '/../routes/fast');
 const mongoose = require('mongoose');
 const PORT = process.env.PORT || 5000;
 mongoose.connect(process.env.MONGO_LOC || 'mongodb://localhost/db');
 
-app.use('/api', router);
+app.use('/api', pedal);
+app.use('/api', motor);
+app.use('/api', fast);
 
 app.listen(PORT, () => process.stdout.write('server lisening on: ' + PORT + '\n'));
